+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 - Breaking: Resolver caches are now stored in PostgreSQL. See [docs/build.md](./docs/build.md) for prerequisite instructions. (#271)
-- Fix: SevenTV emotes now resolve correctly.
+- Fix: SevenTV emotes now resolve correctly. (#281)
+- Dev: Improve BetterTTV emote tests. (#282)
+- Minor: BetterTTV cache key changed from plural to singular form. (#282)
 
 ## 1.2.3